+++ forge/src/main/java/net/mehvahdjukaar/supplementaries/mixins/forge/ChatMixin.java
@@ -2,9 +2,6 @@ package net.mehvahdjukaar.supplementaries.mixins.forge;
 
 import net.mehvahdjukaar.supplementaries.client.renderers.forge.PicklePlayer;
 import net.minecraft.client.gui.screens.ChatScreen;
-import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.world.item.Items;
-import net.minecraftforge.common.extensions.IForgeItem;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -13,10 +10,9 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 @Mixin(ChatScreen.class)
 public abstract class ChatMixin {
 
-    @Inject(method = "handleChatInput", at = @At("HEAD"), cancellable = true)
+    @Inject(method = "handleChatInput", at = @At("HEAD"))
     private void interceptMessage(String string, boolean bl, CallbackInfoReturnable<Boolean> callback) {
         PicklePlayer.onChatEvent(string);
-
     }
 
 }